Modification to logout handler
and login handler
I'm trying to make it send back the userDetails on loginSucess
